requirements.txt
Django
django-admin-sortable2
environs
django-tinymce
pillow
requests
lxml
beautifulsoup4